Server/requirements.txt
customtkinter
pystray
Pillow
darkdetect
